feat: extract into start-validator.ts
