+++ clients/js/cmds/aptos.ts
@@ -8,6 +8,7 @@ import { callEntryFunc, deriveResourceAccount, deriveWrappedAssetAddress } from 
 import { config } from '../config';
 import { NETWORKS } from "../networks";
 import { evm_address, hex } from "../consts";
+import { runCommand, validator_args } from '../start-validator';
 
 type Network = "MAINNET" | "TESTNET" | "DEVNET"
 
@@ -41,14 +42,6 @@ const named_addresses = {
   describe: "Named addresses in the format addr1=0x0,addr2=0x1,...",
   type: "string",
   require: false
-} as const;
-
-const validator_args = {
-  alias: "a",
-  type: "string",
-  array: true,
-  default: [],
-  describe: "Additional args to validator",
 } as const;
 
 // TODO(csongor): this could be useful elsewhere
@@ -430,11 +423,4 @@ function serializePackage(p: Package): PackageBCS {
     bytecodes: serializedModules,
     codeHash
   }
-}
-
-function runCommand(baseCmd: string, args: readonly string[]) {
-  const args_string = args.map(a => `"${a}"`).join(" ");
-  const cmd = `${baseCmd} ${args_string}`;
-  console.log("\x1b[33m%s\x1b[0m", cmd);
-  spawnSync(cmd, { shell: true, stdio: "inherit" });
 }
